renaming package *.collections as *.chainable
