+++ src/main/java/com/github/martinsawicki/chainable/ChainableQueue.java
@@ -1,4 +1,4 @@
-package com.github.martinsawicki.collections;
+package com.github.martinsawicki.chainable;
 
 import com.github.martinsawicki.annotation.Experimental;
