+++ src/main/java/com/github/martinsawicki/chainable/Chainables.java
@@ -2,7 +2,7 @@
  * Copyright (c) Martin Sawicki. All rights reserved.
  * Licensed under the MIT License. See LICENSE in the project root for license information.
  */
-package com.github.martinsawicki.collections;
+package com.github.martinsawicki.chainable;
 
 import java.text.StringCharacterIterator;
 import java.util.ArrayList;
